test(web): split component tests into individual files
