+++ apps/web/src/components/coming-soon.test.tsx
@@ -1,0 +1,14 @@
+import { render, screen } from "@testing-library/react";
+import { ComingSoon } from "@web/components/coming-soon";
+
+describe("ComingSoon", () => {
+  it("should render ComingSoon label", () => {
+    render(
+      <ComingSoon>
+        <span>Trends</span>
+      </ComingSoon>,
+    );
+
+    expect(screen.getByText("Coming Soon")).toBeInTheDocument();
+  });
+});
