+++ apps/web/src/components/structured-data.test.tsx
@@ -1,0 +1,26 @@
+import { render } from "@testing-library/react";
+import { StructuredData } from "@web/components/structured-data";
+import { vi } from "vitest";
+
+vi.mock("next/script", () => ({
+  __esModule: true,
+  default: (props: Record<string, unknown>) => <script {...props} />,
+}));
+
+describe("StructuredData", () => {
+  it("should inject JSON-LD script", () => {
+    const data = {
+      "@context": "https://schema.org",
+      "@type": "Organization",
+      name: "SG Cars Trends",
+    };
+
+    const { container } = render(<StructuredData data={data} />);
+    const script = container.querySelector(
+      "script#structured-data",
+    ) as HTMLScriptElement;
+
+    expect(script).toBeTruthy();
+    expect(script.textContent).toContain("SG Cars Trends");
+  });
+});
